# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Project Hope API
  version: 1.5.0
  description: Tenant-bound API for self-hosted charity operations.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
  - tokenAuth: []
paths:
...
  /pilot-applications/:
    post:
      security: []
      summary: Apply to the Founding 10 pilot
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PilotApplicationRequest' }
      responses:
        '202': { description: Application accepted with a generic privacy-preserving response }
        '400': { description: Invalid input or missing consent }
        '429': { description: Submission rate limit reached }
  /pilot-applications/verify/:
    post:
      security: []
      summary: Confirm a pilot application email using an expiring signed token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, maxLength: 2048 }
      responses:
        '200': { description: Email confirmed }
        '400': { description: Token invalid or expired }
        '429': { description: Verification rate limit reached }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: sessionid }
    tokenAuth: { type: apiKey, in: header, name: Authorization, description: "Use Token <key>." }
    clientHeaders: { type: apiKey, in: header, name: X-Project-Hope-Client }
    clientSecret: { type: apiKey, in: header, name: X-Project-Hope-Secret }
...
  schemas:
    PilotApplicationRequest:
      type: object
      additionalProperties: false
      required: [contact_name, email, organization_name, team_size, primary_need, plan_interest, consent_to_contact]
      properties:
        contact_name: { type: string, maxLength: 160 }
        email: { type: string, format: email, maxLength: 254 }
        organization_name: { type: string, maxLength: 200 }
        website: { type: string, format: uri, maxLength: 500 }
        country_or_region: { type: string, maxLength: 120 }
        team_size: { type: string, enum: ['1', '2-5', '6-20', '21-50', '51+'] }
        primary_need: { type: string, enum: [operations, volunteers, grants, communications, impact, accessibility, other] }
        plan_interest: { type: string, enum: [community, founding_partner, network] }
        notes: { type: string, maxLength: 2000 }
        consent_to_contact: { type: boolean, const: true }
        source: { type: string, enum: [website, linkedin, partner, referral, direct, other] }
        utm_source: { type: string, maxLength: 120 }
        utm_medium: { type: string, maxLength: 120 }
        utm_campaign: { type: string, maxLength: 160 }
        referrer: { type: string, format: uri, maxLength: 500 }
        company_website: { type: string, maxLength: 500, description: Bot-detection honeypot; real clients leave this empty. }